Modification fonctionnement read write de player et enemy
Les fonctions Read et Write du player et de l'enemy ne traitent plus la dimension des données dans le paquet, celle-ci étant fixe.

+++ main/include/enemy.hpp
@@ -31,7 +31,7 @@ struct compRotation
 	uint16_t a;
 	uint16_t b;
 	uint16_t c;
-	uint16_t i;
+	uint8_t i;
 };
 
 class Enemy : GameObject
